services/auth-service/pyproject.toml
[project]
name = "bomberman-auth-service"
version = "0.1.0"
description = "Authentication service for Bomberman Online"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt>=4.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "redis>=5.0.0",
    "aioprometheus>=23.12.0",
    "python-multipart>=0.0.7",
    "email-validator>=2.1.0",
    "python-json-logger>=3.3.0",
    "jinja2>=3.1.0",
    "alembic>=1.13.1",
    "typer>=0.15.4"
] 
